update maven pom file
- add internal and snapshot repository for SIM artifacts
Artifacts produced by SIM will be uploaded to an external maven repo to allow other projects to depend on SIM (for example project Larkc now depends on sim-instrumentation).

- change scope to "provided" for aspectjtools and aspectjweaver
aspectjtools and aspectjweaver are not used at compile time but only at runtime and are not used as classpath libs but as tools.